+++ tests/Feature/Media/Fakes/FileFieldManager.php
@@ -5,7 +5,6 @@ declare(strict_types = 1);
 namespace Thinktomorrow\Chief\Tests\Feature\Media\Fakes;
 
 use Thinktomorrow\Chief\Fields\Fields;
-use Thinktomorrow\Chief\Media\MediaType;
 use Thinktomorrow\Chief\Pages\PageManager;
 use Thinktomorrow\Chief\Fields\Types\FileField;
 
@@ -15,6 +14,7 @@ class FileFieldManager extends PageManager
     {
         return new Fields([
             FileField::make('fake-file'),
+            FileField::make('file-on-other-disk')->storageDisk('secondMediaDisk'),
         ]);
     }
 }